update hydra provider and init hydra instance
