+++ packages/mesh-provider/src/hydra/constants/index.ts
@@ -1,0 +1,11 @@
+export const HYDRA_STATUS = {
+  IDLE: "IDLE",
+  DISCONNECTED: "DISCONNECTED",
+  CONNECTING: "CONNECTING",
+  CONNECTED: "CONNECTED",
+  INITIALIZING: "INITIALIZING",
+  OPEN: "OPEN",
+  CLOSED: "CLOSED",
+  FANOUT_POSSIBLE: "FANOUT_POSSIBLE",
+  FINAL: "FINAL",
+} as const;
